requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
coverage==4.4.1
dj-database-url==0.4.2
Django==1.11.3
django-nose==1.4.4
djangorestframework==3.6.3
Faker==0.7.17
Flask==0.12.2
gevent==1.1.1
greenlet==0.4.12
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
locustio==0.7.5
MarkupSafe==1.0
msgpack-python==0.4.8
nose==1.3.7
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
